[Bronze V] Title: 이상한 기호, Time: 128 ms, Memory: 14204 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -7,8 +7,8 @@ public class Main{
         StringTokenizer st = new StringTokenizer(br.readLine());
         int a = Integer.parseInt(st.nextToken());
         int b = Integer.parseInt(st.nextToken());
-        int cal1 = a+b;
-        int cal2 = a-b;
+        long cal1 = a+b;
+        long cal2 = a-b;
         bw.write(cal1*cal2+"");
         bw.flush();
         bw.close();
